Move db files to own folder
