+++ server/config/passport.js
@@ -4,8 +4,8 @@ var FacebookStrategy = require('passport-facebook').Strategy;
 
 
 /****** REQUIRE DATABASE ******/
-var db = require('./db');
-
+var db = require('../db/config.js');
+var User = require('../db/userSchema.js');
 /****************** PASSPORT CONFIG ***************/
 
 passport.use(new FacebookStrategy({
@@ -15,7 +15,7 @@ passport.use(new FacebookStrategy({
   profileFields: ['id', 'name', 'picture.type(large)', 'email', 'gender']
 },
   function(accessToken, refreshToken, profile, done) {
-    db.User.findOrCreateUser(profile, function(error, user) {
+    User.findOrCreateUser(profile, function(error, user) {
       if (error) {
         return done(error);
       } else {
@@ -34,20 +34,4 @@ passport.deserializeUser(function(user, done) {
   done(null, user);
 });
 
-
 module.exports = passport;
-
-
-/********************** NOTES **************************/
-
-/*
-app.post('/login',
-  passport.authenticate('local', { successRedirect: '/',
-                                   failureRedirect: '/login',
-                                   failureFlash: 'Invalid username or password.' }));
-- Login in valid to / not valid to login
-
-
-{ failureFlash: 'Invalid username or password.' });
-
-*/
