+++ server/db/config.js
@@ -1,0 +1,5 @@
+var mongoose = require('mongoose');
+
+mongoose.connect('mongodb://localhost/voraciousscroll');
+
+module.exports = mongoose;
